+++ internal/bloblang/query/methods_test.go
@@ -733,6 +733,14 @@ func TestMethods(t *testing.T) {
 			),
 			output: `5020219685658847592`,
 		},
+		"check md5 hash": {
+			input: methods(
+				literalFn("hello world"),
+				method("hash", "md5"),
+				method("encode", "hex"),
+			),
+			output: `5eb63bbbe01eeed093cb22bb8f5acdc3`,
+		},
 		"check hex encode": {
 			input: methods(
 				literalFn("hello world"),
@@ -1847,6 +1855,24 @@ func TestMethods(t *testing.T) {
 		"check round down": {
 			input:  methods(literalFn(5.3), method("round")),
 			output: 5.0,
+		},
+		"check replace_many string": {
+			input: methods(literalFn("<i>hello</i> <b>world</b>"), method("replace_many", []interface{}{
+				"<b>", "BOLD",
+				"</b>", "!BOLD",
+				"<i>", "ITA",
+				"</i>", "!ITA",
+			})),
+			output: "ITAhello!ITA BOLDworld!BOLD",
+		},
+		"check replace_many bytes": {
+			input: methods(literalFn([]byte("<i>hello</i> <b>world</b>")), method("replace_many", []interface{}{
+				"<b>", "BOLD",
+				"</b>", "!BOLD",
+				"<i>", "ITA",
+				"</i>", "!ITA",
+			})),
+			output: []byte("ITAhello!ITA BOLDworld!BOLD"),
 		},
 	}